fix(test): two assertions that only held on sqlite
Both failed the moment the suite ran on its canonical connection
(MySQL, arovolife_test) and had been passing on sqlite by accident.

AnalyticsTest inserted a consent row without `doc_hash_sha256`. The
column is declared nullable in the Blueprint and only made
`BINARY(32) NOT NULL` by a MySQL-only ALTER, so sqlite accepted the row.
All three production writers set the hash; the fixture now matches them.

RedeemPointsIsolationTest compared a `SUM()` result to an int. MySQL
returns it as a string through PDO where sqlite returns an int, so the
strict comparison failed. Cast at the assertion.

No production behaviour changes.

+++ app/tests/Modules/Commerce/RedeemPointsIsolationTest.php
@@ -62,5 +62,6 @@ it('RPI-03: the balance check refuses an overdraw', function () {
     expect(fn () => $service->redeem(1, 101, 1, 'too many'))
         ->toThrow(RuntimeException::class, 'the balance is 100');
 
-    expect(RedeemPointEntry::where('distributor_id', 1)->sum('points'))->toBe(100);
+    // MySQL returns SUM() as a string through PDO where sqlite returns an int.
+    expect((int) RedeemPointEntry::where('distributor_id', 1)->sum('points'))->toBe(100);
 });
